+++ src/main/java/roomescape/model/Reservation.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.model;
 
 public class Reservation {
     private int id;
