+++ src/main/java/roomescape/view/ReservationPage.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -7,6 +7,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 public class ReservationPage {
     @GetMapping("/reservation")
     public String reservationPage() {
-        return "reservation";
+        return ViewNames.RESERVATION.getViewName();
     }
 }
