+++ src/main/java/roomescape/view/ViewNames.java
@@ -1,0 +1,16 @@
+package roomescape.view;
+
+public enum ViewNames {
+    HOME("home"),
+    RESERVATION("reservation");
+
+    private final String viewName;
+
+    ViewNames(String viewName) {
+       this.viewName = viewName;
+    }
+
+    public String getViewName() {
+        return viewName;
+    }
+}
